+++ packages/react-native-sdk/src/components/Livestream/LivestreamControls/HostStartStreamButton.tsx
@@ -13,7 +13,7 @@ import {
 } from 'react-native';
 import { useTheme } from '../../../contexts';
 import { EndBroadcastIcon, StartStreamIcon } from '../../../icons';
-import { SfuModels } from '@stream-io/video-client';
+import { SfuModels, getLogger } from '@stream-io/video-client';
 
 /**
  * Props for the HostStartStreamButton component.
@@ -79,7 +79,8 @@ export const HostStartStreamButton = ({
         onStartStreamHandler();
       }
     } catch (error) {
-      console.error('Error starting livestream', error);
+      const logger = getLogger(['HostStartStreamButton']);
+      logger('error', 'Error starting livestream', error);
     }
   };
 
@@ -101,7 +102,8 @@ export const HostStartStreamButton = ({
         onEndStreamHandler();
       }
     } catch (error) {
-      console.error('Error stopping livestream', error);
+      const logger = getLogger(['HostStartStreamButton']);
+      logger('error', 'Error stopping livestream', error);
     }
   };