+++ packages/react-native-sdk/src/utils/StreamVideoRN/index.ts
@@ -75,8 +75,10 @@ export class StreamVideoRN {
       return;
     }
     this.config.push = pushConfig;
+    const foregroundServiceChannel =
+      this.config.foregroundService.android.channel;
     // After getting the config we should setup callkeep events, firebase handler asap to handle incoming calls from a dead state
-    setupFirebaseHandlerAndroid(pushConfig);
+    setupFirebaseHandlerAndroid(pushConfig, foregroundServiceChannel);
     // setup ios handler for non-voip push notifications asap
     setupRemoteNotificationsHandleriOS(pushConfig);
   }