+++ packages/react-native-sdk/src/utils/getAndroidDefaultRingtoneUrl.ts
@@ -1,3 +1,4 @@
+import { getLogger } from '@stream-io/video-client';
 import { NativeModules, Platform } from 'react-native';
 
 export async function getAndroidDefaultRingtoneUrl(): Promise<
@@ -11,7 +12,8 @@ export async function getAndroidDefaultRingtoneUrl(): Promise<
       await NativeModules.StreamVideoReactNative?.getDefaultRingtoneUrl();
     return url;
   } catch (e) {
-    console.info(e);
+    const logger = getLogger(['getAndroidDefaultRingtoneUrl']);
+    logger('warn', 'Failed to get default ringtone from native module', e);
   }
 
   return undefined;